Moved input types to main namespace
