+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/AlertQueryTimeMode.cs
@@ -1,9 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Azure.AI.MetricsAdvisor.Models;
 using Azure.Core;
 
-namespace Azure.AI.MetricsAdvisor.Models
+namespace Azure.AI.MetricsAdvisor
 {
     /// <summary>
     /// Specifies which time property of an <see cref="AnomalyAlert"/> will be used to filter results
