+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/FeedbackDimensionFilter.cs
@@ -3,13 +3,15 @@
 
 using System;
 using System.Collections.Generic;
+using Azure.AI.MetricsAdvisor.Models;
 using Azure.Core;
 
-namespace Azure.AI.MetricsAdvisor.Models
+namespace Azure.AI.MetricsAdvisor
 {
     /// <summary>
     /// Filters the result of feedback-related operations.
     /// </summary>
+    [CodeGenModel("FeedbackDimensionFilter")]
     public partial class FeedbackDimensionFilter
     {
         /// <summary>
