+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/GetAlertsOptions.cs
@@ -2,8 +2,9 @@
 // Licensed under the MIT License.
 
 using System;
+using Azure.AI.MetricsAdvisor.Models;
 
-namespace Azure.AI.MetricsAdvisor.Models
+namespace Azure.AI.MetricsAdvisor
 {
     /// <summary>
     /// The set of options that can be specified when calling <see cref="MetricsAdvisorClient.GetAlerts"/> or
