+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/GetDataFeedsOptions.cs
@@ -1,23 +1,27 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System.Threading;
+using Azure.AI.MetricsAdvisor.Administration;
 
-namespace Azure.AI.MetricsAdvisor.Models
+namespace Azure.AI.MetricsAdvisor
 {
     /// <summary>
-    /// The set of options that can be specified when calling <see cref="MetricsAdvisorClient.GetIncidents(string, string, GetIncidentsForAlertOptions, CancellationToken)"/>
-    /// or <see cref="MetricsAdvisorClient.GetIncidentsAsync(string, string, GetIncidentsForAlertOptions, CancellationToken)"/>
-    /// to configure the behavior of the request.
+    /// The set of options that can be specified when calling <see cref="MetricsAdvisorAdministrationClient.GetDataFeeds"/>
+    /// or <see cref="MetricsAdvisorAdministrationClient.GetDataFeedsAsync"/> to configure the behavior of the request.
     /// </summary>
-    public class GetIncidentsForAlertOptions
+    public class GetDataFeedsOptions
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="GetIncidentsForAlertOptions"/> class.
+        /// Initializes a new instance of the <see cref="GetDataFeedsOptions"/> class.
         /// </summary>
-        public GetIncidentsForAlertOptions()
+        public GetDataFeedsOptions()
         {
         }
+
+        /// <summary>
+        /// Optional filters, such as filtering by name or by status.
+        /// </summary>
+        public GetDataFeedsFilter GetDataFeedsFilter { get; set; }
 
         /// <summary>
         /// If set, skips the first set of items returned. This property specifies the amount of items to
