+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/GetDimensionValuesOptions.cs
@@ -1,27 +1,25 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using Azure.AI.MetricsAdvisor.Administration;
-
-namespace Azure.AI.MetricsAdvisor.Models
+namespace Azure.AI.MetricsAdvisor
 {
     /// <summary>
-    /// The set of options that can be specified when calling <see cref="MetricsAdvisorAdministrationClient.GetDataFeeds"/>
-    /// or <see cref="MetricsAdvisorAdministrationClient.GetDataFeedsAsync"/> to configure the behavior of the request.
+    /// The set of options that can be specified when calling <see cref="MetricsAdvisorClient.GetDimensionValues"/>
+    /// or <see cref="MetricsAdvisorClient.GetDimensionValuesAsync"/> to configure the behavior of the request.
     /// </summary>
-    public class GetDataFeedsOptions
+    public class GetDimensionValuesOptions
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="GetDataFeedsOptions"/> class.
+        /// Initializes a new instance of the <see cref="GetDimensionValuesOptions"/> class.
         /// </summary>
-        public GetDataFeedsOptions()
+        public GetDimensionValuesOptions()
         {
         }
 
         /// <summary>
-        /// Optional filters, such as filtering by name or by status.
+        /// The dimension value to be filtered.
         /// </summary>
-        public GetDataFeedsFilter GetDataFeedsFilter { get; set; }
+        public string DimensionValueToFilter { get; set; }
 
         /// <summary>
         /// If set, skips the first set of items returned. This property specifies the amount of items to
