+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/GetHooksOptions.cs
@@ -1,22 +1,20 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System.Threading;
+using Azure.AI.MetricsAdvisor.Administration;
 
-namespace Azure.AI.MetricsAdvisor.Models
+namespace Azure.AI.MetricsAdvisor
 {
     /// <summary>
-    /// The set of options that can be specified when calling <see cref="MetricsAdvisorClient.GetAnomalies(string, string, GetAnomaliesForAlertOptions, CancellationToken)"/>
-    /// or <see cref="MetricsAdvisorClient.GetAnomaliesAsync(string, string, GetAnomaliesForAlertOptions, CancellationToken)"/> to configure the behavior of the request.
+    /// The set of options that can be specified when calling <see cref="MetricsAdvisorAdministrationClient.GetHooks"/> or
+    /// <see cref="MetricsAdvisorAdministrationClient.GetHooksAsync"/> to configure the behavior of the request.
     /// </summary>
-    public class GetAnomaliesForAlertOptions
+    public class GetHooksOptions
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="GetAnomaliesForAlertOptions"/> class.
+        /// The name of the hook to filter on.
         /// </summary>
-        public GetAnomaliesForAlertOptions()
-        {
-        }
+        public string HookNameFilter { get; set; }
 
         /// <summary>
         /// If set, skips the first set of items returned. This property specifies the amount of items to
